requirements.txt
jupyter
numpy
pandas
matplotlib
seaborn
torch
transformers
datasets
python-dotenv
accelerate
protobuf
sentencepiece
peft
mlflow
streamlit
pydantic
